Move scheduler package
