+++ cmd/restart_test.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/dagu-dev/dagu/internal/dag"
-	"github.com/dagu-dev/dagu/internal/scheduler"
+	"github.com/dagu-dev/dagu/internal/dag/scheduler"
 	"github.com/stretchr/testify/require"
 )
