+++ cmd/retry_test.go
@@ -5,7 +5,7 @@ import (
 	"os"
 	"testing"
 
-	"github.com/dagu-dev/dagu/internal/scheduler"
+	"github.com/dagu-dev/dagu/internal/dag/scheduler"
 	"github.com/stretchr/testify/require"
 )
