+++ internal/agent/reporter.go
@@ -7,8 +7,8 @@ import (
 	"strings"
 
 	"github.com/dagu-dev/dagu/internal/dag"
+	"github.com/dagu-dev/dagu/internal/dag/scheduler"
 	"github.com/dagu-dev/dagu/internal/persistence/model"
-	"github.com/dagu-dev/dagu/internal/scheduler"
 	"github.com/jedib0t/go-pretty/v6/table"
 )
