+++ internal/persistence/jsondb/writer_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/dagu-dev/dagu/internal/persistence/model"
 
 	"github.com/dagu-dev/dagu/internal/dag"
-	"github.com/dagu-dev/dagu/internal/scheduler"
+	"github.com/dagu-dev/dagu/internal/dag/scheduler"
 	"github.com/stretchr/testify/require"
 )
