+++ internal/persistence/model/node.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 
 	"github.com/dagu-dev/dagu/internal/dag"
-	"github.com/dagu-dev/dagu/internal/scheduler"
+	"github.com/dagu-dev/dagu/internal/dag/scheduler"
 	"github.com/dagu-dev/dagu/internal/util"
 )
